+++ serviceTest/CHANGELOG.md
@@ -16,7 +16,7 @@ SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test"
 - add Sam resource type config API
 
 ### Dependency upgrades
-- updated `rawls-model` dependency to `v0.0.251-SNAP`
+- updated `rawls-model` dependency to `v0.0.258-SNAP`
 
 ## 4.4