+++ serviceTest/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 This file documents changes to the `workbench-service-test` library, including notes on how to upgrade to new versions.
 
+## 5.1  
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "5.1-TRAVIS-REPLACE-ME"`
+
+### Dependency upgrades
+- Overwrote some `rawls-model` dependencies to fix the following security findings:
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/15
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/7
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/2
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/18
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/5
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/4
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/6
+  - https://github.com/broadinstitute/workbench-libs/security/dependabot/1
+
+| Dependency                               | Old Version |   New Version |
+|------------------------------------------|:-----------:|--------------:|
+| client-java                              |   19.0.0    | 20.0.0-legacy |
+| google-cloud-kms                         |   2.33.0    |        2.55.0 |
+| net.sourceforge.htmlunit -> org.htmlunit |   2.70.0    |         4.7.0 |
+| selenium-4-1 -> selenium-4-2             |  3.2.12.1   |      3.2.13.0 |
+| akka                                     |   2.6.20    |         2.8.8 |
+| akkaHttp                                 |   10.2.10   |        10.5.3 |
+| jackson                                  |   2.17.1    |        2.18.0 |
+| scalatest                                |   3.2.17    |        3.2.19 |
+
 ## 5.0
 
 SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-service-test" % "5.0-cbb1fb1"`